Bug fix for logMultipleDrinks
With health access disabled tapping the log multiple action would sometimes cause the state of log multiple drinks view to be set to true. When health access was enabled this would lead to the view not being triggered when tapped later.
I think I fixed it but it's hard to reproduce.